feat(logout-usecase): implement active session check

+++ src/application/use-cases/logout.js
@@ -1,3 +1,5 @@
+import { ConflictError } from '../../domain/errors/conflict.js';
+
 class LogoutUseCase {
   #sessionsService = null;
 
@@ -6,6 +8,9 @@ class LogoutUseCase {
   }
 
   async exec() {
+    if (!(await this.#sessionsService.checkAuth())) {
+      throw new ConflictError(ConflictError.NO_SESSION);
+    }
     return this.#sessionsService.destroySession();
   }
 }
